Fixed minor issues
fixed glfw project file (added vulkan .c)
added glfw and glad projects to premake links
added some macroes & comments
deleted Application and Window copy cosntructor
changes to Window constructor
changes to ExitCodes
+ minor changes.

+++ Light Engine/src/Engine/Events/MouseEvents.h
@@ -78,16 +78,16 @@ namespace Light {
 	class MouseScrolledEvent : public Event
 	{
 	private:
-		int8_t m_XOffset;
+		int m_XOffset;
 	public:
-		MouseScrolledEvent(int8_t xOffset): m_XOffset(xOffset) {}
+		MouseScrolledEvent(int xOffset): m_XOffset(xOffset) {}
 
-		inline int8_t GetOffset() { return m_XOffset; }
+		inline int GetOffset() { return m_XOffset; }
 
 		std::string GetLogInfo() const override 
 		{
 			std::stringstream ss;
-			ss << "Mouse scrolled: " << (int)m_XOffset;
+			ss << "Mouse scrolled: " << m_XOffset;
 			return ss.str();
 		}
